fix: waveform.js does not match corresponding path
- replace code in old audio component  file
