+++ test/src/AudioPlayer.js
@@ -1,153 +1,155 @@
-import React, { Component } from "react";
-import ReactHowler from "react-howler";
-import PropTypes from "prop-types";
+import React, { useState, useEffect, useRef } from "react";
+import WaveSurfer from "wavesurfer.js";
+import WaveSurferAsync from "wavesurfer.js-async";
+import regions from "wavesurfer.js/dist/plugin/wavesurfer.regions";
+import "./styles/waveform.css";
+import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
+import {
+  faPlayCircle,
+  faPauseCircle,
+  faBolt,
+} from "@fortawesome/free-solid-svg-icons";
 
-import { PlayButton, Progress, Timer } from "react-soundplayer/components";
+function AudioPlayer({ url }) {
+  const [currentTime, setCurrentTime] = useState("0:00");
+  const [duration, setDuration] = useState("");
+  const [playing, setPlaying] = useState(false);
+  const [speed, setSpeed] = useState(1);
+  const [loading, setLoading] = useState(true);
+  const wavesurfer = useRef(null);
+  const wave = useRef();
 
-const DEFAULT_DURATION = 456.1495; // have to use this become modifying the audio file breaks 2x speed
-const DEFAULT_MP3 =
-  "https://parse-server-ff.s3.amazonaws.com/ae5992f0f5bb1f259bafa41b3771e3bb_call12565815456dwwwwww795896232www-01b59bd3.mp3";
+  const defaultUrl =
+    "https://parse-server-ff.s3.amazonaws.com/ae5992f0f5bb1f259bafa41b3771e3bb_call12565815456dwwwwww795896232www-01b59bd3.mp3";
 
-class AudioPlayer extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      playing: false,
-      currentTime: 0,
-      speedup: false,
-      loadErr: false,
+  useEffect(() => {
+    let interval;
+    setLoading(true);
+    const urlString = url ? url : defaultUrl;
+
+    wavesurfer.current = WaveSurfer.create({
+      barWidth: 1,
+      height: 40,
+      showTime: true,
+      container: "#waveform",
+      progressColor: "#fff",
+      responsive: true,
+      waveColor: "#006",
+      interact: true,
+      backend: "MediaElement",
+      fillParent: true,
+      scrollParent: false,
+      cursorWidth: 3,
+      plugins: [
+        regions.create({
+          regionMinLength: 2,
+          dragSelection: {
+            slop: 5,
+          },
+        }),
+        WaveSurferAsync.create(),
+      ],
+    });
+
+    wavesurfer.current.load(urlString);
+
+    wavesurfer.current.on("ready", () => {
+      setLoading(false);
+      setDuration(formatTime(wavesurfer.current.getDuration()));
+      interval = setInterval(() => {
+        setCurrentTime(formatTime(wavesurfer.current.getCurrentTime()));
+      }, 1000);
+    });
+
+    wavesurfer.current.on("error", (err) => {
+      alert(err);
+    });
+
+    return () => {
+      wavesurfer.current.destroy();
+      clearInterval(interval);
     };
+  }, []);
+
+  // toggle play/pause
+  function playPause() {
+    wavesurfer.current.playPause();
+    setPlaying(!playing);
   }
 
-  seek(secs, play) {
-    if (secs && secs.seek != null) secs = secs.seek();
-    this.player.seek(secs);
-    let toSet = { currentTime: secs };
-    if (play === true) toSet.playing = true;
-    this.setState(toSet);
+  // set playback speed for track (1 is normal speed, 2 is double speed).
+  function setTrackSpeed(rate) {
+    wavesurfer.current.setPlaybackRate(rate);
+    setSpeed(rate);
+    return;
   }
 
-  toggleRate() {
-    let { speedup } = this.state;
-    speedup = !speedup;
-    this.setState({ speedup });
-    this.player._howler.rate(speedup ? 2.0 : 1.0);
+  // format seconds to hours minutes ans seconds
+  function formatTime(time) {
+    // Hours, minutes and seconds
+    var hrs = ~~(time / 3600);
+    var mins = ~~((time % 3600) / 60);
+    var secs = ~~time % 60;
+
+    // Output like "1:01" or "4:03:59" or "123:03:59"
+    var ret = "";
+    if (hrs > 0) {
+      ret += "" + hrs + ":" + (mins < 10 ? "0" : "");
+    }
+    ret += "" + mins + ":" + (secs < 10 ? "0" : "");
+    ret += "" + secs;
+    return ret;
   }
 
-  getState() {
-    let { playing, currentTime } = this.state;
-    return { playing, currentTime };
-  }
+  return (
+    <section className="container">
+      <div className="spectrum-area">
+        <div className="spectrum">
+          <div id="waveform" ref={wave} />
+        </div>
 
-  getSeek() {
-    if (this.playerInterval) clearInterval(this.playerInterval);
-    this.playerInterval = setInterval(() => {
-      let { mp3url } = this.props;
-      if (this.player) {
-        let currentTime = this.player.seek();
-        const duration =
-          mp3url === DEFAULT_MP3 ? DEFAULT_DURATION : this.player.duration();
-        const toSet = { currentTime };
-        if (!this.state.duration && duration != null) {
-          toSet.duration = duration;
-        }
-        if (duration != null) toSet.loadErr = false;
-        if (mp3url === DEFAULT_MP3 && currentTime >= DEFAULT_DURATION) {
-          this.player.stop();
-          toSet.playing = false;
-          currentTime = 0;
-        }
-        this.setState(toSet);
-      }
-    }, 250);
-  }
+        <div className="controls">
+          {/* checks if the track is playing then renders the pause button, if not render the play button  */}
+          {playing ? (
+            <button onClick={playPause} className="controls-button">
+              <FontAwesomeIcon icon={faPauseCircle} size="2x" />
+            </button>
+          ) : (
+            <button onClick={playPause} className="controls-button">
+              <FontAwesomeIcon icon={faPlayCircle} size="2x" />
+            </button>
+          )}
 
-  componentWillUnmount() {
-    if (this.playerInterval) clearTimeout(this.playerInterval);
-  }
-
-  componentDidMount() {
-    this.getSeek();
-  }
-
-  isObject(obj) {
-    return obj instanceof Object || (typeof obj === "object" && obj !== null);
-  }
-
-  render() {
-    const { mp3url } = this.props;
-    let { playing, currentTime, duration, speedup, loadErr } = this.state;
-    if (this.isObject(currentTime)) currentTime = 0;
-    if (mp3url === DEFAULT_MP3) duration = DEFAULT_DURATION;
-    return (
-      <div className="ff-audio">
-        {duration != null ? (
-          <div className="flex flex-center px2 relative z1">
-            <PlayButton
-              playing={playing}
-              onTogglePlay={() => this.setState({ playing: !playing })}
-              className="flex-none h2 mr2 button button-transparent button-grow rounded"
-            />
-            {/* seeking={Boolean}
-                        seekingIcon={ReactElement} */}
-
-            <div className="sb-soundplayer-volume mr2 flex flex-center">
-              <button
-                onClick={() => this.toggleRate()}
-                className="sb-soundplayer-btn sb-soundplayer-volume-btn flex-none h2 button button-transparent button-grow rounded"
-              >
-                <img
-                  className={speedup ? "audio-speedup" : ""}
-                  src="/pane/speedup.svg"
-                  alt="speed-up"
-                  height={35}
-                />
-              </button>
-            </div>
-            <Progress
-              className="flex-auto bg-darken-3 rounded"
-              innerClassName="rounded-left bg-white"
-              value={((currentTime || 0) / (duration || 1)) * 100 || 0}
-              onSeekTrack={(ts) => this.seek(ts * duration)}
-            />
-
-            <Timer
-              className={"timer"}
-              duration={duration} // in seconds
-              currentTime={currentTime != null ? currentTime : 0}
-            />
-          </div>
-        ) : loadErr ? (
-          <div style={{ padding: "5 20px" }}>
-            Unable to load audio: {loadErr}
-          </div>
-        ) : (
-          <div className="progress">
-            <div className="indeterminate" />
-          </div>
-        )}
-        <div>
-          <ReactHowler
-            src={mp3url}
-            playing={playing}
-            loop={false}
-            onLoadError={(id, err) => {
-              console.log("Unable to load media", err);
-              this.setState({
-                loadErr: (err && err.message) || "Startup error",
-              });
-            }}
-            onLoad={() => this.getSeek()}
-            ref={(ref) => (this.player = ref)}
-          />
+          {/* checks the current speed, if speed is one, send two to the finction to increase speed, if not send 1  */}
+          {speed === 1 ? (
+            <button
+              onClick={() => setTrackSpeed(2)}
+              className="controls-button"
+            >
+              2<FontAwesomeIcon icon={faBolt} size="1x" />
+            </button>
+          ) : (
+            <button
+              onClick={() => setTrackSpeed(1)}
+              className="controls-button"
+            >
+              1<FontAwesomeIcon icon={faBolt} size="1x" />
+            </button>
+          )}
         </div>
       </div>
-    );
-  }
-}
 
-AudioPlayer.propTypes = {
-  mp3url: PropTypes.string.isRequired,
-};
+      {loading ? null : (
+        <div className="controls-panel">
+          <div className="timestamp">
+            <p>
+              {currentTime} / {duration}
+            </p>
+          </div>
+        </div>
+      )}
+    </section>
+  );
+}
 
 export default AudioPlayer;
